put non-route functions in their own file and rename for clarity
